Use configured upstream instead of origin/HEAD for prlog and review commands
Branches aren't necessarily based on origin/HEAD, and if they're not,
`prlog` and `review` currently produce bad/useless output. By using a
configured upstream instead, the commands work regardless of the base
branch. Drawback: The commands no longer work with a detached HEAD,
e.g. during a rebase, but the benefits outweigh that.